Water Code Cleanup
