+++ omnimatter_water/locale/en/en.cfg
@@ -3,7 +3,7 @@ water-omnitraction=Water Omnitraction
 mud-omnitraction=Mud Omnitraction
 waste_water=__1__ Omnitraction
 
-water-based-omnitraction=__1__ __2__ omnitraction
+water-based-omnitraction=__1__ __2__ Omnitraction
 
 [recipe-name]
 water-omnitraction=Water
