--- omnimatter_water/prototypes/pyc_aai.lua
@@ -1,3 +1,0 @@
-if mods["pycoalprocessing"] and mods["aai-industry"] then
-
-end
